backend/requirements.txt
django>=4.2.0,<4.3.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
